scripts(pypi): gather very basic metadata about each repo
